requirements.txt
# Core dependencies
apache-airflow==2.7.1
apache-airflow-providers-amazon==8.3.0
apache-airflow-providers-snowflake==5.1.0
dbt-snowflake==1.6.0
great-expectations==0.18.11
pyspark==3.4.1
snowflake-connector-python==3.3.1
snowflake-sqlalchemy==1.5.0

# AWS dependencies
boto3==1.28.38
awscli==1.29.38

# Data processing
pandas==2.1.0
numpy==1.24.3
pyarrow==12.0.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development tools
black==23.7.0
flake8==6.1.0
isort==5.12.0
mypy==1.5.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0 